Add configure.py script for easier distro configuration
Distros usually have different policies for where various package contents should be installed. For instance Fedora installs all user binaries to /usr/bin and has /usr/sbin as a symlink to that while AerynOS and Solus (Arch too I think) don't use /usr/libexec at all and switch those binaries to /usr/lib/some-package-specific-identifier. Some distros follow the hermetic-usr approach and forbid packages from shipping any files outside of /usr and so have alternate /usr destinations for files that typically go in /etc (for example).

Currently packaging is very awkward for distributions that diverge from the "Ubuntu" layout as they need to have downstream patches for the justfiles or custom installation phases in order to get the packages to conform to distro policy which generates a lot of extra work to keep all Cosmic packages in sync.

Instead let's adopt the approach that autotools/meson/cmake use and split out package configuration to an earlier step, and then reuse that configuration for the build and install phases. Do this by adding a configure.py script which accepts common distribution options and which then writes a `.distro_config.env` file. The justfile can then use this .env for the various build settings and installation locations, with default settings for if the .env file is not present (allowing running configure.py to be optional).

By doing this we allow packagers to create common macros which can be shared and used by all cosmic packages and which should be less error-prone than having to keep a set of patches in sync.